+++ weblate/api/views.py
@@ -275,6 +275,7 @@ class DownloadViewSet(viewsets.ReadOnlyModelViewSet):
 class WeblateViewSet(DownloadViewSet):
     """Allow to skip content negotiation for certain requests."""
 
+    @transaction.atomic
     def repository_operation(
         self, request: Request, obj, project: Project, operation: str
     ):
@@ -1875,9 +1876,11 @@ class TranslationViewSet(MultipleFieldViewSet, DestroyModelMixin):
         if not (can_upload := user.has_perm("upload.perform", obj)):
             self.permission_denied(
                 request,
-                can_upload.reason
-                if isinstance(can_upload, PermissionResult)
-                else "Insufficient privileges for uploading.",
+                (
+                    can_upload.reason
+                    if isinstance(can_upload, PermissionResult)
+                    else "Insufficient privileges for uploading."
+                ),
             )
 
         serializer = UploadRequestSerializer(data=request.data)
@@ -2253,33 +2256,45 @@ class UnitViewSet(viewsets.ReadOnlyModelViewSet, UpdateModelMixin, DestroyModelM
         methods=["post"],
         request=CommentSerializer,
     )
-    @action(detail=True, methods=["post"], serializer_class=CommentSerializer)
+    @extend_schema(
+        description="Return a list of comments on the unit.",
+        methods=["get"],
+        responses={200: CommentSerializer(many=True)},
+    )
+    @action(detail=True, methods=["get", "post"], serializer_class=CommentSerializer)
     def comments(self, request, *args, **kwargs):
-        """Add a new comment to a unit."""
+        """List comments on a unit or add a new comment to a unit."""
         unit = self.get_object()
         user = request.user
 
-        if not user.has_perm("comment.add", unit.translation):
-            self.permission_denied(request)
+        if request.method == "POST":
+            if not user.has_perm("comment.add", unit.translation):
+                self.permission_denied(request)
 
-        serializer = CommentSerializer(
-            data=request.data,
-            context={
-                "unit": unit,
-                "request": request,
-            },
-        )
-        serializer.is_valid(raise_exception=True)
+            serializer = CommentSerializer(
+                data=request.data,
+                context={
+                    "unit": unit,
+                    "request": request,
+                },
+            )
+            serializer.is_valid(raise_exception=True)
 
-        timestamp = serializer.validated_data.get("timestamp")
-        user_email = serializer.validated_data.get("user_email")
-        if (timestamp is not None or user_email is not None) and (
-            not user.has_perm("project.edit", unit.translation.component.project)
-        ):
-            self.permission_denied(request)
+            timestamp = serializer.validated_data.get("timestamp")
+            user_email = serializer.validated_data.get("user_email")
+            if (timestamp is not None or user_email is not None) and (
+                not user.has_perm("project.edit", unit.translation.component.project)
+            ):
+                self.permission_denied(request)
 
-        serializer.save()
-        return Response(serializer.data, status=HTTP_201_CREATED)
+            serializer.save()
+            return Response(serializer.data, status=HTTP_201_CREATED)
+
+        # If user doesn't have access to the unit, get_object will raise 404.
+        qs = unit.comment_set.all().order_by("id")
+        page = self.paginate_queryset(qs)
+        serializer = CommentSerializer(page, context={"request": request}, many=True)
+        return self.get_paginated_response(serializer.data)
 
 
 @extend_schema_view(