+++ weblate/settings_docker.py
@@ -797,7 +797,6 @@ INSTALLED_APPS = [
     # Third party Django modules
     "social_django",
     "crispy_forms",
-    "crispy_bootstrap3",
     "crispy_bootstrap5",
     "compressor",
     "rest_framework",
@@ -1088,8 +1087,8 @@ SIMPLIFY_LANGUAGES = get_env_bool("WEBLATE_SIMPLIFY_LANGUAGES", True)
 DEFAULT_PAGE_LIMIT = get_env_int("WEBLATE_DEFAULT_PAGE_LIMIT", 100)
 
 # Render forms using bootstrap
-CRISPY_ALLOWED_TEMPLATE_PACKS = ["bootstrap3", "bootstrap5"]
-CRISPY_TEMPLATE_PACK = "bootstrap3"
+CRISPY_ALLOWED_TEMPLATE_PACKS = "bootstrap5"
+CRISPY_TEMPLATE_PACK = "bootstrap5"
 
 # List of quality checks
 CHECK_LIST = [